+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 
 #### Builtin
 
+- **Breaking Change**: remove ``Array#forEach`` method due to unused
 - **Breaking Change**: throw ``InvalidOperationError`` when call ``Module#_func`` method within user-defined completer
 - ``shctl module`` sub-command now finds and prints full path of specified modules